Add session management
Re-using the same gssapi security context with every request causes
issues with the authentication towards the IPA server. This
implementation will authenticate once and then use the session endpoint
to continue the authentication.

Any request that is successfully authenticated will set a session cookie
in the response that is valid in the /ipa/session/json endpoint.
HTTPClient automatically handles the cookies within the object.

In IPA 4.4 the correct way of authenticating is to issue a request to
the /ipa/session/login_kerberos endpoint, this is not available in 4.2.0
though.